requirements.txt
# ─── Project Roomy — Python Dependencies ────────────────────────────────────

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# LLM providers
anthropic>=0.25.0
openai>=1.30.0
httpx>=0.27.0          # HTTP client (also used for Ollama)

# Database
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Telegram bot
python-telegram-bot>=21.0

# Environment
python-dotenv>=1.0.0

# Redis (Alfred's session memory)
redis>=5.0.0

# Utilities
aiofiles>=23.0.0       # Async file handling (for image uploads in Phase 3)

# ── Dev only (remove in production) ──────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx                  # also used for test client
